Rewrote gamerules system.
Now when a gamerule is set we can track the lastvalue set and also if the value is changed run a callback that adjusts whatever is needed for the gamerule change to work. Some of the functions added are reset_gamerule( rulename ) which resets a gamerule to the original value that it had when the server started. set_gamerule_for_match( rulename, rulevalue ) which sets the gamerule for the match and calls the gamerule's callback if it exists. Lastly set_gamerule_for_next_matches( rulename, rulevalue, number_of_matches ) was added which will be used to set a gamerule for a specific number of matches or forever before resetting back to startup value.